requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
PGPy==0.6.0
playwright==1.36.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pyee==9.0.4
requests==2.31.0
requests-toolbelt==1.0.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
